clean up remove function
