+++ client/ayon_unreal/plugins/load/load_layout_existing.py
@@ -18,9 +18,7 @@ class ExistingLayoutLoader(plugin.LayoutLoader):
     """
 
     label = "Load Layout on Existing Scene"
-
     delete_unmatched_assets = True
-    loaded_layout_dir = "{folder[path]}/{product[name]}"
 
     @classmethod
     def apply_settings(cls, project_settings):
@@ -33,6 +31,10 @@ class ExistingLayoutLoader(plugin.LayoutLoader):
         cls.loaded_layout_dir = (
             project_settings["unreal"].get(
                 "loaded_layout_dir", cls.loaded_layout_dir)
+        )
+        cls.remove_loaded_assets = (
+            project_settings["unreal"].get(
+                "remove_loaded_assets", cls.remove_loaded_assets)
         )
 
     def _spawn_actor(self, obj, lasset, sequence):
@@ -321,11 +323,3 @@ class ExistingLayoutLoader(plugin.LayoutLoader):
             container, repre_entity, loaded_assets=loaded_assets)
 
         unreal.EditorLevelLibrary.save_current_level()
-
-    def remove(self, container):
-        parent_path = Path(container["namespace"])
-        container_name = container["container_name"]
-        if unreal.EditorAssetLibrary.does_asset_exist(
-            f"{parent_path}/{container_name}"):
-                unreal.EditorAssetLibrary.delete_asset(
-                    f"{parent_path}/{container_name}")
